refactor(mcp): split middleware.rs into per-concern submodules
Promotes mcp/middleware.rs (2,245 lines) to a directory module.
Middleware implementations move to dedicated files:
- logging.rs (LoggingMiddleware, LogLevel)
- validation.rs (ValidationMiddleware)
- performance.rs (PerformanceMiddleware)
- auth.rs (AuthenticationMiddleware + JWT/OAuth types)
- rate_limit.rs (RateLimitMiddleware)
- config.rs (all MiddlewareConfig structs)

mod.rs retains core types (MiddlewareContext, MiddlewareResult,
McpMiddleware trait, MiddlewareChain) and the integration tests.
All types remain pub-re-exported at the same path.
