+++ apps/things3-cli/src/mcp/middleware/mod.rs
@@ -1,17 +1,11 @@
 //! MCP Middleware system for cross-cutting concerns
 
 use crate::mcp::{CallToolRequest, CallToolResult, McpError, McpResult};
-use governor::clock::DefaultClock;
-use governor::{state::keyed::DefaultKeyedStateStore, Quota, RateLimiter};
 #[allow(unused_imports)]
 use jsonwebtoken::{decode, encode, Algorithm, DecodingKey, EncodingKey, Header, Validation};
-use nonzero_ext::nonzero;
-use serde::{Deserialize, Serialize};
 use serde_json::Value;
-use std::collections::HashMap;
 use std::sync::Arc;
 use std::time::{Duration, Instant};
-use thiserror::Error;
 
 /// Middleware execution context
 #[derive(Debug, Clone)]
@@ -228,851 +222,24 @@ impl Default for MiddlewareChain {
     }
 }
 
-/// Built-in logging middleware
-pub struct LoggingMiddleware {
-    level: LogLevel,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum LogLevel {
-    Debug,
-    Info,
-    Warn,
-    Error,
-}
-
-impl LoggingMiddleware {
-    /// Create a new logging middleware
-    #[must_use]
-    pub fn new(level: LogLevel) -> Self {
-        Self { level }
-    }
-
-    /// Create with debug level
-    #[must_use]
-    pub fn debug() -> Self {
-        Self::new(LogLevel::Debug)
-    }
-
-    /// Create with info level
-    #[must_use]
-    pub fn info() -> Self {
-        Self::new(LogLevel::Info)
-    }
-
-    /// Create with warn level
-    #[must_use]
-    pub fn warn() -> Self {
-        Self::new(LogLevel::Warn)
-    }
-
-    /// Create with error level
-    #[must_use]
-    pub fn error() -> Self {
-        Self::new(LogLevel::Error)
-    }
-
-    fn should_log(&self, level: LogLevel) -> bool {
-        matches!(
-            (self.level, level),
-            (LogLevel::Debug, _)
-                | (
-                    LogLevel::Info,
-                    LogLevel::Info | LogLevel::Warn | LogLevel::Error
-                )
-                | (LogLevel::Warn, LogLevel::Warn | LogLevel::Error)
-                | (LogLevel::Error, LogLevel::Error)
-        )
-    }
-
-    fn log(&self, level: LogLevel, message: &str) {
-        if self.should_log(level) {
-            match level {
-                LogLevel::Debug => println!("[DEBUG] {message}"),
-                LogLevel::Info => println!("[INFO] {message}"),
-                LogLevel::Warn => println!("[WARN] {message}"),
-                LogLevel::Error => println!("[ERROR] {message}"),
-            }
-        }
-    }
-}
-
-#[async_trait::async_trait]
-impl McpMiddleware for LoggingMiddleware {
-    fn name(&self) -> &'static str {
-        "logging"
-    }
-
-    fn priority(&self) -> i32 {
-        100 // Low priority to run early
-    }
-
-    async fn before_request(
-        &self,
-        request: &CallToolRequest,
-        context: &mut MiddlewareContext,
-    ) -> McpResult<MiddlewareResult> {
-        self.log(
-            LogLevel::Info,
-            &format!(
-                "Request started: {} (ID: {})",
-                request.name, context.request_id
-            ),
-        );
-        Ok(MiddlewareResult::Continue)
-    }
-
-    async fn after_request(
-        &self,
-        request: &CallToolRequest,
-        response: &mut CallToolResult,
-        context: &mut MiddlewareContext,
-    ) -> McpResult<MiddlewareResult> {
-        let elapsed = context.elapsed();
-        let status = if response.is_error {
-            "ERROR"
-        } else {
-            "SUCCESS"
-        };
-
-        self.log(
-            LogLevel::Info,
-            &format!(
-                "Request completed: {} (ID: {}) - {} in {:?}",
-                request.name, context.request_id, status, elapsed
-            ),
-        );
-        Ok(MiddlewareResult::Continue)
-    }
-
-    async fn on_error(
-        &self,
-        request: &CallToolRequest,
-        error: &McpError,
-        context: &mut MiddlewareContext,
-    ) -> McpResult<MiddlewareResult> {
-        self.log(
-            LogLevel::Error,
-            &format!(
-                "Request failed: {} (ID: {}) - {}",
-                request.name, context.request_id, error
-            ),
-        );
-        Ok(MiddlewareResult::Continue)
-    }
-}
-
-/// Built-in validation middleware
-pub struct ValidationMiddleware {
-    strict_mode: bool,
-}
-
-impl ValidationMiddleware {
-    /// Create a new validation middleware
-    #[must_use]
-    pub fn new(strict_mode: bool) -> Self {
-        Self { strict_mode }
-    }
-
-    /// Create with strict mode enabled
-    #[must_use]
-    pub fn strict() -> Self {
-        Self::new(true)
-    }
-
-    /// Create with strict mode disabled
-    #[must_use]
-    pub fn lenient() -> Self {
-        Self::new(false)
-    }
-
-    fn validate_request(&self, request: &CallToolRequest) -> McpResult<()> {
-        // Basic validation
-        if request.name.is_empty() {
-            return Err(McpError::validation_error("Tool name cannot be empty"));
-        }
-
-        // Validate tool name format (alphanumeric and underscores only)
-        if !request
-            .name
-            .chars()
-            .all(|c| c.is_alphanumeric() || c == '_')
-        {
-            return Err(McpError::validation_error(
-                "Tool name must contain only alphanumeric characters and underscores",
-            ));
-        }
-
-        // In strict mode, validate arguments structure
-        if self.strict_mode {
-            if let Some(args) = &request.arguments {
-                if !args.is_object() {
-                    return Err(McpError::validation_error(
-                        "Arguments must be a JSON object",
-                    ));
-                }
-            }
-        }
-
-        Ok(())
-    }
-}
-
-#[async_trait::async_trait]
-impl McpMiddleware for ValidationMiddleware {
-    fn name(&self) -> &'static str {
-        "validation"
-    }
-
-    fn priority(&self) -> i32 {
-        50 // Medium priority
-    }
-
-    async fn before_request(
-        &self,
-        request: &CallToolRequest,
-        context: &mut MiddlewareContext,
-    ) -> McpResult<MiddlewareResult> {
-        if let Err(error) = self.validate_request(request) {
-            context.set_metadata(
-                "validation_error".to_string(),
-                serde_json::Value::String(error.to_string()),
-            );
-            return Ok(MiddlewareResult::Error(error));
-        }
-
-        context.set_metadata("validated".to_string(), serde_json::Value::Bool(true));
-        Ok(MiddlewareResult::Continue)
-    }
-}
-
-/// Built-in performance monitoring middleware
-pub struct PerformanceMiddleware {
-    slow_request_threshold: Duration,
-}
-
-impl PerformanceMiddleware {
-    /// Create a new performance middleware
-    #[must_use]
-    pub fn new(slow_request_threshold: Duration) -> Self {
-        Self {
-            slow_request_threshold,
-        }
-    }
-
-    /// Create with default threshold (1 second)
-    #[must_use]
-    pub fn create_default() -> Self {
-        Self::new(Duration::from_secs(1))
-    }
-
-    /// Create with custom threshold
-    #[must_use]
-    pub fn with_threshold(threshold: Duration) -> Self {
-        Self::new(threshold)
-    }
-}
-
-#[async_trait::async_trait]
-impl McpMiddleware for PerformanceMiddleware {
-    fn name(&self) -> &'static str {
-        "performance"
-    }
-
-    fn priority(&self) -> i32 {
-        200 // High priority to run late
-    }
-
-    async fn after_request(
-        &self,
-        request: &CallToolRequest,
-        _response: &mut CallToolResult,
-        context: &mut MiddlewareContext,
-    ) -> McpResult<MiddlewareResult> {
-        let elapsed = context.elapsed();
-
-        // Record performance metrics
-        context.set_metadata(
-            "duration_ms".to_string(),
-            serde_json::Value::Number(serde_json::Number::from(
-                u64::try_from(elapsed.as_millis()).unwrap_or(u64::MAX),
-            )),
-        );
-
-        context.set_metadata(
-            "is_slow".to_string(),
-            serde_json::Value::Bool(elapsed > self.slow_request_threshold),
-        );
-
-        // Log slow requests
-        if elapsed > self.slow_request_threshold {
-            println!(
-                "[PERF] Slow request detected: {} took {:?} (threshold: {:?})",
-                request.name, elapsed, self.slow_request_threshold
-            );
-        }
-
-        Ok(MiddlewareResult::Continue)
-    }
-}
-
-/// Authentication middleware for API key and OAuth 2.0 support
-pub struct AuthenticationMiddleware {
-    api_keys: HashMap<String, ApiKeyInfo>,
-    jwt_secret: String,
-    #[allow(dead_code)]
-    oauth_config: Option<OAuthConfig>,
-    require_auth: bool,
-}
-
-#[derive(Debug, Clone)]
-pub struct ApiKeyInfo {
-    pub key_id: String,
-    pub permissions: Vec<String>,
-    pub expires_at: Option<chrono::DateTime<chrono::Utc>>,
-}
-
-#[derive(Debug, Clone)]
-pub struct OAuthConfig {
-    pub client_id: String,
-    pub client_secret: String,
-    pub token_endpoint: String,
-    pub scope: Vec<String>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct JwtClaims {
-    pub sub: String, // Subject (user ID)
-    pub exp: usize,  // Expiration time
-    pub iat: usize,  // Issued at
-    pub permissions: Vec<String>,
-}
-
-impl AuthenticationMiddleware {
-    /// Create a new authentication middleware
-    #[must_use]
-    pub fn new(api_keys: HashMap<String, ApiKeyInfo>, jwt_secret: String) -> Self {
-        Self {
-            api_keys,
-            jwt_secret,
-            oauth_config: None,
-            require_auth: true,
-        }
-    }
-
-    /// Create with OAuth 2.0 support
-    #[must_use]
-    pub fn with_oauth(
-        api_keys: HashMap<String, ApiKeyInfo>,
-        jwt_secret: String,
-        oauth_config: OAuthConfig,
-    ) -> Self {
-        Self {
-            api_keys,
-            jwt_secret,
-            oauth_config: Some(oauth_config),
-            require_auth: true,
-        }
-    }
-
-    /// Create without requiring authentication (for testing)
-    #[must_use]
-    pub fn permissive() -> Self {
-        Self {
-            api_keys: HashMap::new(),
-            jwt_secret: "test-secret".to_string(),
-            oauth_config: None,
-            require_auth: false,
-        }
-    }
-
-    /// Extract API key from request headers or arguments
-    fn extract_api_key(request: &CallToolRequest) -> Option<String> {
-        // Check if API key is in request arguments
-        if let Some(args) = &request.arguments {
-            if let Some(api_key) = args.get("api_key").and_then(|v| v.as_str()) {
-                return Some(api_key.to_string());
-            }
-        }
-        None
-    }
-
-    /// Extract JWT token from request headers or arguments
-    fn extract_jwt_token(request: &CallToolRequest) -> Option<String> {
-        // Check if JWT token is in request arguments
-        if let Some(args) = &request.arguments {
-            if let Some(token) = args.get("jwt_token").and_then(|v| v.as_str()) {
-                return Some(token.to_string());
-            }
-        }
-        None
-    }
-
-    /// Validate API key
-    fn validate_api_key(&self, api_key: &str) -> McpResult<ApiKeyInfo> {
-        self.api_keys
-            .get(api_key)
-            .cloned()
-            .ok_or_else(|| McpError::validation_error("Invalid API key"))
-    }
-
-    /// Validate JWT token
-    fn validate_jwt_token(&self, token: &str) -> McpResult<JwtClaims> {
-        let validation = Validation::new(Algorithm::HS256);
-        let key = DecodingKey::from_secret(self.jwt_secret.as_ref());
-
-        let token_data = decode::<JwtClaims>(token, &key, &validation)
-            .map_err(|_| McpError::validation_error("Invalid JWT token"))?;
-
-        // Check if token is expired
-        let now = chrono::Utc::now().timestamp().try_into().unwrap_or(0);
-        if token_data.claims.exp < now {
-            return Err(McpError::validation_error("JWT token has expired"));
-        }
-
-        Ok(token_data.claims)
-    }
-
-    /// Generate JWT token for testing
-    ///
-    /// # Panics
-    /// Panics if JWT encoding fails
-    #[cfg(test)]
-    #[must_use]
-    pub fn generate_test_jwt(&self, user_id: &str, permissions: Vec<String>) -> String {
-        #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-        let now = chrono::Utc::now().timestamp() as usize;
-        let claims = JwtClaims {
-            sub: user_id.to_string(),
-            exp: now + 3600, // 1 hour
-            iat: now,
-            permissions,
-        };
-
-        let header = Header::new(Algorithm::HS256);
-        let key = EncodingKey::from_secret(self.jwt_secret.as_ref());
-        encode(&header, &claims, &key).unwrap()
-    }
-}
-
-#[async_trait::async_trait]
-impl McpMiddleware for AuthenticationMiddleware {
-    fn name(&self) -> &'static str {
-        "authentication"
-    }
-
-    fn priority(&self) -> i32 {
-        10 // High priority to run early
-    }
-
-    async fn before_request(
-        &self,
-        request: &CallToolRequest,
-        context: &mut MiddlewareContext,
-    ) -> McpResult<MiddlewareResult> {
-        if !self.require_auth {
-            context.set_metadata("auth_required".to_string(), Value::Bool(false));
-            return Ok(MiddlewareResult::Continue);
-        }
-
-        // Try API key authentication first
-        if let Some(api_key) = Self::extract_api_key(request) {
-            if let Ok(api_key_info) = self.validate_api_key(&api_key) {
-                context.set_metadata(
-                    "auth_type".to_string(),
-                    Value::String("api_key".to_string()),
-                );
-                context.set_metadata(
-                    "auth_key_id".to_string(),
-                    Value::String(api_key_info.key_id),
-                );
-                context.set_metadata(
-                    "auth_permissions".to_string(),
-                    serde_json::to_value(api_key_info.permissions).unwrap_or(Value::Array(vec![])),
-                );
-                context.set_metadata("auth_required".to_string(), Value::Bool(true));
-                return Ok(MiddlewareResult::Continue);
-            }
-            // API key failed, try JWT
-        }
-
-        // Try JWT authentication
-        if let Some(jwt_token) = Self::extract_jwt_token(request) {
-            if let Ok(claims) = self.validate_jwt_token(&jwt_token) {
-                context.set_metadata("auth_type".to_string(), Value::String("jwt".to_string()));
-                context.set_metadata("auth_user_id".to_string(), Value::String(claims.sub));
-                context.set_metadata(
-                    "auth_permissions".to_string(),
-                    serde_json::to_value(claims.permissions).unwrap_or(Value::Array(vec![])),
-                );
-                context.set_metadata("auth_required".to_string(), Value::Bool(true));
-                return Ok(MiddlewareResult::Continue);
-            }
-            // JWT failed
-        }
-
-        // No valid authentication found
-        let error_result = CallToolResult {
-            content: vec![crate::mcp::Content::Text {
-                text: "Authentication required. Please provide a valid API key or JWT token."
-                    .to_string(),
-            }],
-            is_error: true,
-        };
-
-        Ok(MiddlewareResult::Stop(error_result))
-    }
-}
-
-/// Rate limiting middleware with per-client limits
-pub struct RateLimitMiddleware {
-    rate_limiter: Arc<RateLimiter<String, DefaultKeyedStateStore<String>, DefaultClock>>,
-    default_limit: u32,
-    #[allow(dead_code)]
-    burst_limit: u32,
-}
-
-impl RateLimitMiddleware {
-    /// Create a new rate limiting middleware
-    #[must_use]
-    pub fn new(requests_per_minute: u32, burst_limit: u32) -> Self {
-        let quota = Quota::per_minute(nonzero!(60u32)); // Use a constant for now
-        let rate_limiter = Arc::new(RateLimiter::keyed(quota));
-
-        Self {
-            rate_limiter,
-            default_limit: requests_per_minute,
-            burst_limit,
-        }
-    }
-
-    /// Create with custom limits
-    #[must_use]
-    pub fn with_limits(requests_per_minute: u32, burst_limit: u32) -> Self {
-        Self::new(requests_per_minute, burst_limit)
-    }
-
-    /// Create with default limits (60 requests per minute, burst of 10)
-    #[allow(clippy::should_implement_trait)]
-    #[must_use]
-    pub fn default() -> Self {
-        Self::new(60, 10)
-    }
-
-    /// Extract client identifier from request
-    fn extract_client_id(request: &CallToolRequest, context: &MiddlewareContext) -> String {
-        // Try to get from authentication context first
-        if let Some(auth_key_id) = context.get_metadata("auth_key_id").and_then(|v| v.as_str()) {
-            return format!("api_key:{auth_key_id}");
-        }
-
-        if let Some(auth_user_id) = context
-            .get_metadata("auth_user_id")
-            .and_then(|v| v.as_str())
-        {
-            return format!("jwt:{auth_user_id}");
-        }
-
-        // Fallback to request-based identifier
-        if let Some(args) = &request.arguments {
-            if let Some(client_id) = args.get("client_id").and_then(|v| v.as_str()) {
-                return format!("client:{client_id}");
-            }
-        }
-
-        // Use request ID as fallback
-        format!("request:{}", context.request_id)
-    }
-
-    /// Check if request is within rate limits
-    fn check_rate_limit(&self, client_id: &str) -> bool {
-        self.rate_limiter.check_key(&client_id.to_string()).is_ok()
-    }
-
-    /// Get remaining requests for client
-    fn get_remaining_requests(&self, _client_id: &str) -> u32 {
-        // This is a simplified implementation
-        // In a real implementation, you'd want to track remaining requests more precisely
-        self.default_limit
-    }
-}
-
-#[async_trait::async_trait]
-impl McpMiddleware for RateLimitMiddleware {
-    fn name(&self) -> &'static str {
-        "rate_limiting"
-    }
-
-    fn priority(&self) -> i32 {
-        20 // Run after authentication but before other middleware
-    }
-
-    async fn before_request(
-        &self,
-        request: &CallToolRequest,
-        context: &mut MiddlewareContext,
-    ) -> McpResult<MiddlewareResult> {
-        let client_id = Self::extract_client_id(request, context);
-
-        if !self.check_rate_limit(&client_id) {
-            let error_result = CallToolResult {
-                content: vec![crate::mcp::Content::Text {
-                    text: format!(
-                        "Rate limit exceeded. Limit: {} requests per minute. Please try again later.",
-                        self.default_limit
-                    ),
-                }],
-                is_error: true,
-            };
-
-            context.set_metadata("rate_limited".to_string(), Value::Bool(true));
-            context.set_metadata("rate_limit_client_id".to_string(), Value::String(client_id));
-
-            return Ok(MiddlewareResult::Stop(error_result));
-        }
-
-        let remaining = self.get_remaining_requests(&client_id);
-        context.set_metadata(
-            "rate_limit_remaining".to_string(),
-            Value::Number(serde_json::Number::from(remaining)),
-        );
-        context.set_metadata("rate_limit_client_id".to_string(), Value::String(client_id));
-
-        Ok(MiddlewareResult::Continue)
-    }
-}
-
-/// Security configuration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SecurityConfig {
-    /// Authentication configuration
-    pub authentication: AuthenticationConfig,
-    /// Rate limiting configuration
-    pub rate_limiting: RateLimitingConfig,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AuthenticationConfig {
-    /// Enable authentication middleware
-    pub enabled: bool,
-    /// Require authentication for all requests
-    pub require_auth: bool,
-    /// JWT secret for token validation
-    pub jwt_secret: String,
-    /// API keys configuration
-    pub api_keys: Vec<ApiKeyConfig>,
-    /// OAuth 2.0 configuration
-    pub oauth: Option<OAuth2Config>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ApiKeyConfig {
-    /// API key value
-    pub key: String,
-    /// Key identifier
-    pub key_id: String,
-    /// Permissions for this key
-    pub permissions: Vec<String>,
-    /// Optional expiration date
-    pub expires_at: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct OAuth2Config {
-    /// OAuth client ID
-    pub client_id: String,
-    /// OAuth client secret
-    pub client_secret: String,
-    /// Token endpoint URL
-    pub token_endpoint: String,
-    /// Required scopes
-    pub scopes: Vec<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RateLimitingConfig {
-    /// Enable rate limiting middleware
-    pub enabled: bool,
-    /// Requests per minute limit
-    pub requests_per_minute: u32,
-    /// Burst limit for short bursts
-    pub burst_limit: u32,
-    /// Custom limits per client type
-    pub custom_limits: Option<HashMap<String, u32>>,
-}
-
-impl Default for SecurityConfig {
-    fn default() -> Self {
-        Self {
-            authentication: AuthenticationConfig {
-                enabled: true,
-                require_auth: false, // Start with auth disabled for easier development
-                jwt_secret: "your-secret-key-change-this-in-production".to_string(),
-                api_keys: vec![],
-                oauth: None,
-            },
-            rate_limiting: RateLimitingConfig {
-                enabled: true,
-                requests_per_minute: 60,
-                burst_limit: 10,
-                custom_limits: None,
-            },
-        }
-    }
-}
-
-/// Middleware configuration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct MiddlewareConfig {
-    /// Logging configuration
-    pub logging: LoggingConfig,
-    /// Validation configuration
-    pub validation: ValidationConfig,
-    /// Performance monitoring configuration
-    pub performance: PerformanceConfig,
-    /// Security configuration
-    pub security: SecurityConfig,
-}
-
-/// Logging middleware configuration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct LoggingConfig {
-    /// Enable logging middleware
-    pub enabled: bool,
-    /// Log level for logging middleware
-    pub level: String,
-}
-
-/// Validation middleware configuration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ValidationConfig {
-    /// Enable validation middleware
-    pub enabled: bool,
-    /// Use strict validation mode
-    pub strict_mode: bool,
-}
-
-/// Performance monitoring configuration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PerformanceConfig {
-    /// Enable performance monitoring
-    pub enabled: bool,
-    /// Slow request threshold in milliseconds
-    pub slow_request_threshold_ms: u64,
-}
-
-impl Default for MiddlewareConfig {
-    fn default() -> Self {
-        Self {
-            logging: LoggingConfig {
-                enabled: true,
-                level: "info".to_string(),
-            },
-            validation: ValidationConfig {
-                enabled: true,
-                strict_mode: false,
-            },
-            performance: PerformanceConfig {
-                enabled: true,
-                slow_request_threshold_ms: 1000,
-            },
-            security: SecurityConfig::default(),
-        }
-    }
-}
-
-impl MiddlewareConfig {
-    /// Create a new middleware configuration
-    #[must_use]
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Build a middleware chain from this configuration
-    #[must_use]
-    pub fn build_chain(self) -> MiddlewareChain {
-        let mut chain = MiddlewareChain::new();
-
-        // Security middleware (highest priority)
-        if self.security.authentication.enabled {
-            let api_keys: HashMap<String, ApiKeyInfo> = self
-                .security
-                .authentication
-                .api_keys
-                .into_iter()
-                .map(|config| {
-                    let expires_at = config.expires_at.and_then(|date_str| {
-                        chrono::DateTime::parse_from_rfc3339(&date_str)
-                            .ok()
-                            .map(|dt| dt.with_timezone(&chrono::Utc))
-                    });
-
-                    let api_key_info = ApiKeyInfo {
-                        key_id: config.key_id,
-                        permissions: config.permissions,
-                        expires_at,
-                    };
-
-                    (config.key, api_key_info)
-                })
-                .collect();
-
-            let auth_middleware = if self.security.authentication.require_auth {
-                if let Some(oauth_config) = self.security.authentication.oauth {
-                    let oauth = OAuthConfig {
-                        client_id: oauth_config.client_id,
-                        client_secret: oauth_config.client_secret,
-                        token_endpoint: oauth_config.token_endpoint,
-                        scope: oauth_config.scopes,
-                    };
-                    AuthenticationMiddleware::with_oauth(
-                        api_keys,
-                        self.security.authentication.jwt_secret,
-                        oauth,
-                    )
-                } else {
-                    AuthenticationMiddleware::new(api_keys, self.security.authentication.jwt_secret)
-                }
-            } else {
-                AuthenticationMiddleware::permissive()
-            };
-
-            chain = chain.add_middleware(auth_middleware);
-        }
-
-        if self.security.rate_limiting.enabled {
-            let rate_limit_middleware = RateLimitMiddleware::with_limits(
-                self.security.rate_limiting.requests_per_minute,
-                self.security.rate_limiting.burst_limit,
-            );
-            chain = chain.add_middleware(rate_limit_middleware);
-        }
-
-        if self.logging.enabled {
-            let log_level = match self.logging.level.to_lowercase().as_str() {
-                "debug" => LogLevel::Debug,
-                "warn" => LogLevel::Warn,
-                "error" => LogLevel::Error,
-                _ => LogLevel::Info,
-            };
-            chain = chain.add_middleware(LoggingMiddleware::new(log_level));
-        }
-
-        if self.validation.enabled {
-            chain = chain.add_middleware(ValidationMiddleware::new(self.validation.strict_mode));
-        }
-
-        if self.performance.enabled {
-            let threshold = Duration::from_millis(self.performance.slow_request_threshold_ms);
-            chain = chain.add_middleware(PerformanceMiddleware::with_threshold(threshold));
-        }
-
-        chain
-    }
-}
-
-/// Middleware-specific errors
-#[derive(Error, Debug)]
+mod auth;
+mod config;
+mod logging;
+mod performance;
+mod rate_limit;
+mod validation;
+
+pub use auth::{ApiKeyInfo, AuthenticationMiddleware, JwtClaims, OAuthConfig};
+pub use config::{
+    ApiKeyConfig, AuthenticationConfig, LoggingConfig, MiddlewareConfig, OAuth2Config,
+    PerformanceConfig, RateLimitingConfig, SecurityConfig, ValidationConfig,
+};
+pub use logging::{LogLevel, LoggingMiddleware};
+pub use performance::PerformanceMiddleware;
+pub use rate_limit::RateLimitMiddleware;
+pub use validation::ValidationMiddleware;
+
+#[derive(Debug, thiserror::Error)]
 pub enum MiddlewareError {
     #[error("Middleware execution failed: {message}")]
     ExecutionFailed { message: String },
@@ -1094,6 +261,7 @@ impl From<MiddlewareError> for McpError {
 mod tests {
     use super::*;
     use crate::mcp::Content;
+    use std::collections::HashMap;
 
     struct TestMiddleware {
         priority: i32,
